Moving serial scheduling model to a subdir.
