+++ iree/hal/dylib/CMakeLists.txt
@@ -39,7 +39,7 @@ iree_cc_library(
     ::dylib_device
     iree::hal::device_info
     iree::hal::driver
-    iree::hal::host::serial_scheduling_model
+    iree::hal::host::serial::serial_scheduling_model
   PUBLIC
 )
