+++ iree/hal/dylib/dylib_driver.cc
@@ -18,7 +18,7 @@
 
 #include "iree/hal/device_info.h"
 #include "iree/hal/dylib/dylib_device.h"
-#include "iree/hal/host/serial_scheduling_model.h"
+#include "iree/hal/host/serial/serial_scheduling_model.h"
 
 namespace iree {
 namespace hal {
