+++ iree/hal/host/BUILD
@@ -25,10 +25,10 @@ cc_library(
     srcs = ["async_command_queue.cc"],
     hdrs = ["async_command_queue.h"],
     deps = [
-        ":serial_submission_queue",
         "//iree/base:status",
         "//iree/base:tracing",
         "//iree/hal:command_queue",
+        "//iree/hal/host/serial:serial_submission_queue",
         "//iree/hal:semaphore",
         "@com_google_absl//absl/base:core_headers",
         "@com_google_absl//absl/synchronization",
@@ -40,11 +40,11 @@ cc_test(
     srcs = ["async_command_queue_test.cc"],
     deps = [
         ":async_command_queue",
-        ":serial_submission_queue",
         "//iree/base:status",
         "//iree/base:status_matchers",
         "//iree/base:time",
         "//iree/hal:command_queue",
+        "//iree/hal/host/serial:serial_submission_queue",
         "//iree/hal/testing:mock_command_buffer",
         "//iree/hal/testing:mock_command_queue",
         "//iree/testing:gtest_main",
@@ -187,57 +187,5 @@ cc_library(
     hdrs = ["scheduling_model.h"],
     deps = [
         "//iree/hal:command_queue",
-    ],
-)
-
-cc_library(
-    name = "serial_command_processor",
-    srcs = ["serial_command_processor.cc"],
-    hdrs = ["serial_command_processor.h"],
-    deps = [
-        ":host_descriptor_set",
-        ":host_executable",
-        ":host_executable_layout",
-        "//iree/base:source_location",
-        "//iree/base:status",
-        "//iree/base:tracing",
-        "//iree/hal:command_buffer",
-        "@com_google_absl//absl/container:inlined_vector",
-    ],
-)
-
-cc_library(
-    name = "serial_scheduling_model",
-    srcs = ["serial_scheduling_model.cc"],
-    hdrs = ["serial_scheduling_model.h"],
-    deps = [
-        ":async_command_queue",
-        ":host_event",
-        ":host_semaphore",
-        ":inproc_command_buffer",
-        ":scheduling_model",
-        ":serial_command_processor",
-        ":serial_submission_queue",
-        "//iree/base:memory",
-        "//iree/base:source_location",
-        "//iree/base:status",
-        "//iree/base:tracing",
-        "@com_google_absl//absl/container:inlined_vector",
-    ],
-)
-
-cc_library(
-    name = "serial_submission_queue",
-    srcs = ["serial_submission_queue.cc"],
-    hdrs = ["serial_submission_queue.h"],
-    deps = [
-        ":host_semaphore",
-        "//iree/base:intrusive_list",
-        "//iree/base:status",
-        "//iree/base:tracing",
-        "//iree/hal:command_queue",
-        "@com_google_absl//absl/base:core_headers",
-        "@com_google_absl//absl/container:inlined_vector",
-        "@com_google_absl//absl/synchronization",
     ],
 )
