+++ iree/hal/host/CMakeLists.txt
@@ -22,12 +22,12 @@ iree_cc_library(
   SRCS
     "async_command_queue.cc"
   DEPS
-    ::serial_submission_queue
     absl::core_headers
     absl::synchronization
     iree::base::status
     iree::base::tracing
     iree::hal::command_queue
+    iree::hal::host::serial::serial_submission_queue
     iree::hal::semaphore
   PUBLIC
 )
@@ -39,13 +39,13 @@ iree_cc_test(
     "async_command_queue_test.cc"
   DEPS
     ::async_command_queue
-    ::serial_submission_queue
     absl::memory
     absl::time
     iree::base::status
     iree::base::status_matchers
     iree::base::time
     iree::hal::command_queue
+    iree::hal::host::serial::serial_submission_queue
     iree::hal::testing::mock_command_buffer
     iree::hal::testing::mock_command_queue
     iree::testing::gtest_main
@@ -213,67 +213,6 @@ iree_cc_library(
   HDRS
     "scheduling_model.h"
   DEPS
-    iree::hal::command_queue
-  PUBLIC
-)
-
-iree_cc_library(
-  NAME
-    serial_command_processor
-  HDRS
-    "serial_command_processor.h"
-  SRCS
-    "serial_command_processor.cc"
-  DEPS
-    ::host_descriptor_set
-    ::host_executable
-    ::host_executable_layout
-    absl::inlined_vector
-    iree::base::source_location
-    iree::base::status
-    iree::base::tracing
-    iree::hal::command_buffer
-  PUBLIC
-)
-
-iree_cc_library(
-  NAME
-    serial_scheduling_model
-  HDRS
-    "serial_scheduling_model.h"
-  SRCS
-    "serial_scheduling_model.cc"
-  DEPS
-    ::async_command_queue
-    ::host_event
-    ::host_semaphore
-    ::inproc_command_buffer
-    ::scheduling_model
-    ::serial_command_processor
-    ::serial_submission_queue
-    absl::inlined_vector
-    iree::base::memory
-    iree::base::source_location
-    iree::base::status
-    iree::base::tracing
-  PUBLIC
-)
-
-iree_cc_library(
-  NAME
-    serial_submission_queue
-  HDRS
-    "serial_submission_queue.h"
-  SRCS
-    "serial_submission_queue.cc"
-  DEPS
-    ::host_semaphore
-    absl::core_headers
-    absl::inlined_vector
-    absl::synchronization
-    iree::base::intrusive_list
-    iree::base::status
-    iree::base::tracing
     iree::hal::command_queue
   PUBLIC
 )
