+++ iree/hal/host/serial/serial_command_processor.cc
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "iree/hal/host/serial_command_processor.h"
+#include "iree/hal/host/serial/serial_command_processor.h"
 
 #include "iree/base/source_location.h"
 #include "iree/base/status.h"
