+++ iree/hal/host/serial/serial_scheduling_model.h
@@ -12,8 +12,8 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#ifndef IREE_HAL_HOST_SERIAL_SCHEDULING_MODEL_H_
-#define IREE_HAL_HOST_SERIAL_SCHEDULING_MODEL_H_
+#ifndef IREE_HAL_HOST_SERIAL_SERIAL_SCHEDULING_MODEL_H_
+#define IREE_HAL_HOST_SERIAL_SERIAL_SCHEDULING_MODEL_H_
 
 #include "absl/container/inlined_vector.h"
 #include "iree/base/memory.h"
@@ -61,4 +61,4 @@ class SerialSchedulingModel final : public SchedulingModel {
 }  // namespace hal
 }  // namespace iree
 
-#endif  // IREE_HAL_HOST_SERIAL_SCHEDULING_MODEL_H_
+#endif  // IREE_HAL_HOST_SERIAL_SERIAL_SCHEDULING_MODEL_H_
