+++ iree/hal/llvmjit/BUILD
@@ -38,7 +38,7 @@ cc_library(
         ":llvmjit_device",
         "//iree/hal:device_info",
         "//iree/hal:driver",
-        "//iree/hal/host:serial_scheduling_model",
+        "//iree/hal/host/serial:serial_scheduling_model",
         "@llvm-project//llvm:ExecutionEngine",
     ],
 )
