+++ iree/hal/llvmjit/CMakeLists.txt
@@ -40,7 +40,7 @@ iree_cc_library(
     LLVMExecutionEngine
     iree::hal::device_info
     iree::hal::driver
-    iree::hal::host::serial_scheduling_model
+    iree::hal::host::serial::serial_scheduling_model
   PUBLIC
 )
