+++ iree/hal/vmla/BUILD
@@ -103,7 +103,7 @@ cc_library(
         "//iree/base:tracing",
         "//iree/hal:device_info",
         "//iree/hal:driver",
-        "//iree/hal/host:serial_scheduling_model",
+        "//iree/hal/host/serial:serial_scheduling_model",
         "//iree/vm:instance",
         "//iree/vm:module",
     ],
